Refactor: CustomUserDetails 클래스 auth 패키지로 이동
