+++ src/main/java/com/example/hobbyheavy/jwt/JWTUtil.java
@@ -1,6 +1,5 @@
 package com.example.hobbyheavy.jwt;
 
-import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
